Separate applications load generator into its own package
