+++ components/applications-load-gen/.gitignore
@@ -1,0 +1,2 @@
+cmd/applications-load-gen/applications-load-gen
+applications-load-gen
