+++ components/applications-load-gen/cmd/applications-load-gen/applications_load_gen.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/chef/automate/components/applications-load-gen/cmd/applications-load-gen/commands"
+
+func main() {
+	commands.Execute()
+}
