+++ components/applications-load-gen/habitat/default.toml
@@ -1,0 +1,16 @@
+[mlsa]
+accept = false
+
+[service]
+host = "localhost"
+port = 10133
+metrics_port = 20133
+
+[storage]
+database = "chef_applications_service"
+user = "applications"
+
+[tls]
+key_contents = ""
+cert_contents = ""
+root_cert_contents = ""
