+++ components/applications-load-gen/habitat/hooks/init
@@ -1,0 +1,11 @@
+#!{{pkgPathFor "core/bash"}}/bin/bash
+# *** WARNING ***
+# Currently, the Habitat `init` and `health_check` hooks run directly from the main loop
+# of the Habitat supervisor. If these hooks hang or take too long to run, they can block
+# execution of the supervisor.
+#
+# Please put potentially long-running and/or complex operations in the run hook rather
+# than the init hook until the issue[0] has been resolved.
+#
+# [0] https://github.com/habitat-sh/habitat/issues/1973
+#
