+++ components/applications-load-gen/habitat/hooks/run
@@ -1,0 +1,11 @@
+#!{{pkgPathFor "core/bash"}}/bin/bash
+
+set -e
+
+exec 2>&1
+
+# Call the script to block until user accepts the MLSA via the package's config
+{{pkgPathFor "chef/mlsa"}}/bin/accept {{cfg.mlsa.accept}}
+
+# Start our service
+exec applications-load-gen run --profile {{pkg.svc_config_path}}/config.toml
