--- components/applications-service/cmd/applications-load-gen/applications_load_gen.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/chef/automate/components/applications-service/cmd/applications-load-gen/commands"
-
-func main() {
-	commands.Execute()
-}
